Add ctest support
Adds testcase.py, derived from the one used in the clamav tests.

Adds feature to find a clamav installation when the tests are enabled.
It looks for ClamScan, SigTool, ClamBC, and the ClamBC Headers.
The above applications will be useful in future tests to validate that
the compiled signatures are correct. The ClamBC Headers will be required
to compile signatures when running the test suite in the future when the
ClamAV installation includes clambc generated headers (anticipated for
0.104.0), and when this project drops those headers. We presently carry
those headers for 0.103.1.

Adds a `--version` option to clambc-compiler.py, though it is presently
hardcoded to 0.103.1.  In the future, it should get the version from
libclambcc.so if possible, which inherits that value from the project
settings at the top of the main CMakeLists.txt file.

Modified clambc-compiler.py so it will use LD_LIBRARY_PATH to find
libclambcc.so if it can't find it in `../lib`.

Added clamav to the Dockerfile builder image so we can run ctest, but
since ctest currently fails some tests, it is set to ignore failures.