clambc-compiler: Add type hints & fix unused function "genExe"
The genExe() function was using an undefined variable for the clang
version number. I added the ClangLLVM class object as a parameter so I
could just call clangLLVM.getClang() and remove the undefined variable.

While at it, I got distracted and added a bunch of type hints and did a
bit of assorted cleanup.