Fix copyright statement
Compiler was previously adding source code when copyright statement was
in the source because the __Copyright variable was being marked as
internal by built-in optimization pass 'internalize', and removed by built-in
optimizers (likely globalopt, but not verified) because it wasn't used.

Fix this by adding it to the 'internalize-public-api-list' option.