Restructure libclambcc directory to fix unit tests
The unit tests rely on all of the opt pass libraries being located in
the same directory as libclambccommon.so, as specified using
LD_LIBRARY_PATH.

One solution would have been to specify a path for each opt pass library
in LD_LIBRARY_PATH and then search for each library in each of those
LD_LIBRARY_PATH paths. However this felt very clunky.

Instead, this commit solves the problem by placing the source for all
opt passes in the same directory, which will mean that the .so's created
in the build directory will also be co-located and thus using a single
path for LD_LIBRARY_PATH will suffice.

To make present and future copy-pasting easy, I also renamed the
libraries to be camel-case.